requirements.txt
financetoolkit==1.6.4
yfinance==0.2.40
pandas==2.2.2
Riskfolio-Lib